requirements.txt
langchain-core>=0.2.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-groq
streamlit
pandas
python-dotenv
pydantic>=2.0